Refactor of OtaClientStatistics collecting logic (#87)
The previous implementation of OtaClientStatistics collecting logic may cause the statistics not being updated properly or missing updates when grpc status API being called with very short interval.
This PR fixes this problem by refactoring the statistics collecting logic, and improves the performance. The format of OtaStatistics is not changed.