[Fix] cboot.is_switching_boot: only check ota_status and slot_in_use (#231)
This PR lifts the conditions check for is_switching_boot, as long as the ota_status and slot_in_use indicates that we are in finalizing switching boot stage, we do the finalizing.
If ota_status and slot_in_use indicates a successful switching boot, but nvbootctrl shows the slot is not unbootable, only issue a warning for it. (check the following section about nvbootctl for more details).

Check PR description for motivation and more details related to cboot switching boot mechanism.