[Refinement] refactor RetryTaskMap (#232)
This PR refactors RetryTaskMap, which realizes the feature of the new downloading retry strategy.

Design of new RetryTaskMap
New RetryTaskMap comes with two classes:
1. _TaskMap: the executor for each retry round,
2. RetryTaskMap: managing retry rounds, provides map API for starting mapping, and shutdown API to interrupt the execution.

_TaskMap is a threadpool-based task executor for a single retry round. it has a dispatcher thread to dispatching tasks into the threadpool, while exposes a generator instance(returned by calling map method) for upper caller to yield the done task. Done tasks is collected by each task's callback, and put in to a que, while generator instance will yield from the que.
After the all the input tasks is dispatched and finished, it will return a list of failed tasks if any, or None to indicate the success of this retry round.

RetryTaskMap manages retry rounds, it manages instance of _TaskMap for each retry round, yield from the _TaskMap inst for the upper caller. For each retry round, if there are failed tasks from this retry round, RetryTaskMap will cleanup the old _TaskMap instance, and then init and start new instance of _TaskMap, with input as previous failed tasks.
It will continue the above procedure, until all the tasks are finished successfully, or being interrupted due to exceeding retry count, or by custom interruption logic.