fix(boot_control): jetson_common, grub: support detecting and update rootfs indicating string that use device path  (#393)
Currently, the boot control implementations for grub and jetson_* assume that the rootfs indicating string(in the form of root=...) is either using PARTUUID or UUID to specify rootfs or this string is not presented, but not covering the case of directly using device path. This PR fixes this issue and support updating rootfs indicating string that previously was by device path.

NOTE that although we support parsing device path now, we still not support device path with space!