refactor: split slot mount helper as a standalone module, refactor to make the mount/umount more robust (#417)
This PR implements the robust mount/umount for SlotMountHelper. Now otaclient will try its best to ensure the mount point is clear(no other mounts on it) before mount, and for standby slot mount, also ensure the standby slot dev is not mounted before mounting. The failure reason for the mount failure will also be logged.