feat: create_standby.rebuild_mode: file_table at delta_src assisted delta_calculation (#487)
This PR adds feature of using file_table database of delta_source(for current rebuild_mode, the active_slot's file_table) to assist and speed up delta calculation. 
With delta_source file_table presented and using sqlite3 query, we can have a list of files from delta_source with digests that we need in the OTA image, and only scan those files. Previously, we have to fully scan the delta_source to get the files that we can re-use.

With this PR, now `create_standby.DeltaGenerator` has two variants, one is the new `DeltaGenWithFileTable`, and the old `DeltaGenFullDiskScan`.
`ota_core` will check if the delta_source's file_table is presented and valid or not, if available, use `DeltaGenWithFileTable` to calculate delta.

This PR bumps to use simple-sqlite3-orm v0.10.0.