[3.9] fix: replace_root and TopDownCommonShortestPath: DO NOT resolve input paths (#592)
This PR fixes a bug caused by replace_root and TopDownCommonShortestPath try to resolve the input paths.

This will introduce side effect when input path is somehow not an actual dir, but a symlink, on the active slot.
The path will be resolved by looking at current rootfs filesystem, which is unexpected if the input path should be relative to standby slot during delta calculation.

Also, paths resolve is not needed in the first place, so the change here doesn't break the current behavior.