fix: otaproxy cannot commit cache to db when the cache entry already exists (#613)
This PR fixes an issue of OTA cache commit will fail when the cache file is already presented. Now otaproxy will check whether the cache file to save already exists or not, if that is a file, skip overriding the file.

This kind of thing happens when previous OTA is interrupted, and there is a chance the cache file is saved but the cache entry is not committed to the database.