fix!: remove rollback API support (#639)
The rollback API is not used, while exposes a potential DoS vulnerability that can be utilized by attackers that has local network access to the otaclient grpc server listen address due to the request is not signed and verified.

This PR removes the actual rollback grpc API handler, replaced with a dummy handler that responds with failure RECOVERABLE and message rollback support is removed, indicating rollback API is invalid.

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>