fix: otaproxy py313 compatibility: fix cache rotating query (#697)
This PR fixes a bug that only triggered on newer python version. The otaproxy cache rotating query which uses RETURNING statement is actually incorrect, this PR corrects the query.

Previously this bug is not found in test CI due to test setup only covers upto python 3.10(ubuntu 22.04), and the bug is revealed during dev of #695.